requirements.txt
# VectorizerE - Graph-Based Markdown Vectorizer
networkx>=3.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langgraph>=0.2.0
chromadb>=0.4.0
typing-extensions>=4.0
tiktoken>=0.5.0

# VisualizeGraphE - Graph Visualization
pyvis>=0.3.0
matplotlib>=3.7.0

# GraphRAG Bridge utilities
pandas>=2.2.0
pyarrow>=14.0.0
numpy>=1.26.0

# Gmail API / Email Pipeline
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.0

# HTML to Markdown for emails
html2text>=2020.1.16